Breaks up the really large message-handlers tests into a series of smaller test files and fixes all of their lint errors.
